pyproject.toml
[project]
name = "brazilian-stocks"
version = "0.1.0"
description = ""
authors = [
    {name = "Thiago Panini",email = "panini.development@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "boto3 (>=1.36.13,<2.0.0)",
    "boto3-stubs (>=1.36.14,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "brazilian_stocks", from = "src"}]


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.4"
pre-commit = "^4.1.0"
pytest = "^8.3.4"
requests-mock = "^1.12.1"
mypy = "^1.14.1"
pytest-cov = "^6.0.0"
moto = "^5.0.28"


[tool.poetry.group.discovery.dependencies]
jupyter = "^1.1.1"
pandas = "^2.2.3"
investpy = "^1.0.8"
pydantic = "^2.10.6"
yfinance = "^0.2.52"
beautifulsoup4 = "^4.12.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
